+++ e2e/vercel/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @e2e/vercel
 
+## 0.0.64
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.64
+
+## 0.0.63
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.63
+
+## 0.0.62
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.62
+
+## 0.0.61
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.61
+
+## 0.0.60
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.60
+
+## 0.0.59
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @e2e/shared-server@0.0.59
+
 ## 0.0.58
 
 ### Patch Changes